+++ alpha/external-web/app/routes/declare-your-visit.js
@@ -1,5 +1,5 @@
 var router = require('../routes')
-var client = require('../services/eligibility-client')
+var client = require('../services/database/eligibility-client')
 var logger = require('../services/bunyan-logger')
 
 // Require file upload library.
